reogranize package API
